# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColdSpace Solutions API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContactFormInputService =
  (typeof ContactFormInputService)[keyof typeof ContactFormInputService];

export const ContactFormInputService = {
  hvac: "hvac",
  refrigeration: "refrigeration",
  maintenance: "maintenance",
  emergency: "emergency",
  other: "other",
} as const;

export type ContactFormInputLanguage =
  (typeof ContactFormInputLanguage)[keyof typeof ContactFormInputLanguage];

export const ContactFormInputLanguage = {
  en: "en",
  es: "es",
} as const;

export interface ContactFormInput {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  email: string;
  /** @maxLength 20 */
  phone?: string;
  service: ContactFormInputService;
  /** @maxLength 1000 */
  message: string;
  language?: ContactFormInputLanguage;
}

export interface ContactFormResponse {
  success: boolean;
  message: string;
  id?: number;
}

export interface ErrorResponse {
  error: string;
}

export interface ContactItem {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  service: string;
  message: string;
  language: string;
  createdAt: string;
}

export interface ContentItem {
  key: string;
  value?: string | null;
  updatedAt: string;
}

export interface UpdateContentInput {
  value: string | null;
}