# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ColdSpace Solutions API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form operations
  - name: content
    description: Site content operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /contact:
    get:
      operationId: getContacts
      tags: [contact]
      summary: Get all contacts
      description: Retrieves all contact submissions
      responses:
        "200":
          description: A list of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactItem"
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      description: Submits a contact/lead form from the website
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactFormInput"
      responses:
        "201":
          description: Contact form submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactFormResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content:
    get:
      operationId: getContent
      tags: [content]
      summary: Get site content
      description: Retrieves all dynamic site content
      responses:
        "200":
          description: A list of content key-value pairs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
  /content/{key}:
    put:
      operationId: updateContent
      tags: [content]
      summary: Update a content value
      description: Updates the value for a specific content key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentInput"
      responses:
        "200":
          description: Successfully updated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ContactFormInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          maxLength: 20
        service:
          type: string
          enum:
            - hvac
            - refrigeration
            - maintenance
            - emergency
            - other
        message:
          type: string
          maxLength: 1000
        language:
          type: string
          enum: [en, es]
          default: en
      required:
        - name
        - email
        - service
        - message
    ContactFormResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        id:
          type: integer
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ContactItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        service:
          type: string
        message:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - service
        - message
        - language
        - createdAt
    ContentItem:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required:
        - key
        - updatedAt
    UpdateContentInput:
      type: object
      properties:
        value:
          type: string
          nullable: true
      required:
        - value