# lib/api-zod/src/generated/types/contactItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColdSpace Solutions API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactItem {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  service: string;
  message: string;
  language: string;
  createdAt: Date;
}