# lib/api-zod/src/generated/types/contentItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColdSpace Solutions API
 * OpenAPI spec version: 0.1.0
 */

export interface ContentItem {
  key: string;
  value?: string | null;
  updatedAt: Date;
}